+++ crates/base/core/src/code/marker/mod.rs
@@ -8,7 +8,7 @@
 
 mod _reexport; // SYMLINK from /src/code/marker/_reexport_core.rs
 
-mod prim; // Prim, PrimFitPtr
+mod prim; // Prim, PrimFitPtr, PrimIndex
 mod type_marker; // zero-cost generic type markers
 
 crate::structural_mods! { // _mods, _reexports