+++ crates/base/core/src/code/marker/prim.rs
@@ -1,11 +1,12 @@
 // devela_base_core::num::grain::prim
 //
-//! Marker traits for primitive scalars.
+//! Marker traits for primitives.
 //
 // TOC
 // - (trait Sealed)
 // - trait Prim
 // - trait PrimFitPtr
+// - trait PrimIndex
 
 #![expect(private_bounds, reason = "Sealed traits")]
 
@@ -36,7 +37,7 @@ impl_prim![Prim for
     f32, f64,
 ];
 
-/* pointer-width sized */
+/* pointer-width related */
 
 #[doc = crate::_tags!(code mem)]
 /// Primitive value types that fit in pointer-width on supported Rust targets.
@@ -50,3 +51,16 @@ impl_prim![PrimFitPtr for bool, u16, i16];
 impl_prim![PrimFitPtr for u16, u32, i16, i32, f32];
 #[cfg(target_pointer_width = "64")]
 impl_prim![PrimFitPtr for u16, u32, u64, i16, i32, i64, f32, f64];
+
+#[doc = crate::_tags!(code mem num)]
+/// Primitive types that can be used for indexing.
+#[doc = crate::_doc_location!("code/marker")]
+pub trait PrimIndex: crate::PrimUint + PrimFitPtr {}
+
+impl_prim![PrimIndex for u8, usize];
+#[cfg(target_pointer_width = "16")]
+impl_prim![PrimIndex for u16];
+#[cfg(target_pointer_width = "32")]
+impl_prim![PrimIndex for u16, u32];
+#[cfg(target_pointer_width = "64")]
+impl_prim![PrimIndex for u16, u32, u64];