+++ src/code/marker/mod.rs
@@ -18,7 +18,7 @@ crate::structural_mods! { // _mods, _reexports
         pub use super::_reexport_core::*;
         #[doc(inline)]
         pub use devela_base_core::code::marker::{
-            Prim, PrimFitPtr,
+            Prim, PrimFitPtr, PrimIndex,
             type_marker,
         };
     }